fix: js modules remove
