+++ index.html
@@ -136,6 +136,6 @@
       </div>
     </article>
   </template>
-  <script type="module" src="script/courses.js"></script>
+  <script src="script/courses.js"></script>
 </body>
 </html>
