--- script/userInfo.js
@@ -1,13 +1,0 @@
-export const user = {
-  name: 'Иванова Анна Сидоровна',
-  level: 'Профессионал',
-  profession: 'Картограф',
-  courses: 2,
-  birthday: '1990-01-01',
-  location: 'г. Санкт-Петербург',
-  nickname: 'Белка',
-  photo: 'images/profile/belka.svg',
-  phone: '+7 123 456 78 90',
-  email: 'anna@liza-alert.ru',
-  password: 'qwerty'
-}
